refactor: query, command 컨트롤러, 서비스로 분리
