+++ src/backend/chat-service/src/main/java/capstone/chatservice/domain/server/service/command/ServerMessageCommandService.java
@@ -1,20 +1,15 @@
-package capstone.chatservice.domain.server.service;
+package capstone.chatservice.domain.server.service.command;
 
 import capstone.chatservice.domain.server.dto.ServerMessageDto;
 import capstone.chatservice.domain.server.dto.request.ServerMessageCreateRequest;
 import capstone.chatservice.domain.server.dto.request.ServerMessageDeleteRequest;
 import capstone.chatservice.domain.server.dto.request.ServerMessageModifyRequest;
-import org.springframework.data.domain.Page;
 
-public interface ServerMessageService {
+public interface ServerMessageCommandService {
 
     ServerMessageDto save(ServerMessageCreateRequest createRequest);
 
     ServerMessageDto modify(ServerMessageModifyRequest modifyRequest);
 
     ServerMessageDto delete(ServerMessageDeleteRequest deleteRequest);
-
-    Page<ServerMessageDto> getMessages(Long channelId, int page, int size);
-
-    Page<ServerMessageDto> getComments(Long parentId, int page, int size);
 }
