+++ src/backend/chat-service/src/main/java/capstone/chatservice/domain/server/service/query/ServerMessageQueryService.java
@@ -1,0 +1,11 @@
+package capstone.chatservice.domain.server.service.query;
+
+import capstone.chatservice.domain.server.dto.ServerMessageDto;
+import org.springframework.data.domain.Page;
+
+public interface ServerMessageQueryService {
+
+    Page<ServerMessageDto> getMessages(Long channelId, int page, int size);
+
+    Page<ServerMessageDto> getComments(Long parentId, int page, int size);
+}
